// src/core/engine/dash/Dash.d.ts
declare namespace Dash {
// ...
  export interface Bitrate {
    id?: string;
    width?: number;
    height?: number;
    bandwidth?: number;
    scanType?: string;
  }

  export type MediaType = 'video' | 'audio' | 'text' | 'image';

  export class MediaInfo {
    id: string | null;
    index: number | null;
    type: MediaType | null;
    streamInfo: StreamInfo | null;
    representationCount: number;
    labels: string[];
    lang: string | null;
    viewpoint: any | undefined | null;
    accessibility: any[] | null;
    audioChannelConfiguration: any[] | null;
    roles: string[] | null;
    codec: string | null;
    mimeType: string | null;
    contentProtection: any | null;
    isText: boolean;
    KID: any | null;
    bitrateList: Bitrate[];
    supplementalProperties: object;
  }
// ...
  export namespace MediaPlayer {
    export const events: MediaPlayerEvents;
    export const errors: MediaPlayerErrors;
  }

  interface MediaPlayerErrors {
    MANIFEST_LOADER_PARSING_FAILURE_ERROR_CODE: 10;
    MANIFEST_LOADER_LOADING_FAILURE_ERROR_CODE: 11;
    XLINK_LOADER_LOADING_FAILURE_ERROR_CODE: 12;
    SEGMENT_BASE_LOADER_ERROR_CODE: 15;
    TIME_SYNC_FAILED_ERROR_CODE: 16;
    FRAGMENT_LOADER_LOADING_FAILURE_ERROR_CODE: 17;
    FRAGMENT_LOADER_NULL_REQUEST_ERROR_CODE: 18;
    URL_RESOLUTION_FAILED_GENERIC_ERROR_CODE: 19;
    APPEND_ERROR_CODE: 20;
    REMOVE_ERROR_CODE: 21;
    DATA_UPDATE_FAILED_ERROR_CODE: 22;
    CAPABILITY_MEDIASOURCE_ERROR_CODE: 23;
    CAPABILITY_MEDIAKEYS_ERROR_CODE: 24;
    DOWNLOAD_ERROR_ID_MANIFEST_CODE: 25;
    DOWNLOAD_ERROR_ID_SIDX_CODE: 26;
    DOWNLOAD_ERROR_ID_CONTENT_CODE: 27;
    DOWNLOAD_ERROR_ID_INITIALIZATION_CODE: 28;
    DOWNLOAD_ERROR_ID_XLINK_CODE: 29;
    MANIFEST_ERROR_ID_PARSE_CODE: 31;
    MANIFEST_ERROR_ID_NOSTREAMS_CODE: 32;
    TIMED_TEXT_ERROR_ID_PARSE_CODE: 33;
    MANIFEST_ERROR_ID_MULTIPLEXED_CODE: 34;
    MEDIASOURCE_TYPE_UNSUPPORTED_CODE: 35;
    MEDIA_KEYERR_CODE: 100;
    MEDIA_KEYERR_UNKNOWN_CODE: 101;
    MEDIA_KEYERR_CLIENT_CODE: 102;
    MEDIA_KEYERR_SERVICE_CODE: 103;
    MEDIA_KEYERR_OUTPUT_CODE: 104;
    MEDIA_KEYERR_HARDWARECHANGE_CODE: 105;
    MEDIA_KEYERR_DOMAIN_CODE: 106;
    MEDIA_KEY_MESSAGE_ERROR_CODE: 107;
    MEDIA_KEY_MESSAGE_NO_CHALLENGE_ERROR_CODE: 108;
    SERVER_CERTIFICATE_UPDATED_ERROR_CODE: 109;
    KEY_STATUS_CHANGED_EXPIRED_ERROR_CODE: 110;
    MEDIA_KEY_MESSAGE_NO_LICENSE_SERVER_URL_ERROR_CODE: 111;
    KEY_SYSTEM_ACCESS_DENIED_ERROR_CODE: 112;
    KEY_SESSION_CREATED_ERROR_CODE: 113;
    MEDIA_KEY_MESSAGE_LICENSER_ERROR_CODE: 114;
    // MSS errors
    MSS_NO_TFRF_CODE: 200;
    MSS_UNSUPPORTED_CODEC_CODE: 201;
    // Offline errors
    OFFLINE_ERROR: 11000;
    INDEXEDDB_QUOTA_EXCEED_ERROR: 11001;
    INDEXEDDB_INVALID_STATE_ERROR: 11002;
    INDEXEDDB_NOT_READABLE_ERROR: 11003;
    INDEXEDDB_NOT_FOUND_ERROR: 11004;
    INDEXEDDB_NETWORK_ERROR: 11005;
    INDEXEDDB_DATA_ERROR: 11006;
    INDEXEDDB_TRANSACTION_INACTIVE_ERROR: 11007;
    INDEXEDDB_NOT_ALLOWED_ERROR: 11008;
    INDEXEDDB_NOT_SUPPORTED_ERROR: 11009;
    INDEXEDDB_VERSION_ERROR: 11010;
    INDEXEDDB_TIMEOUT_ERROR: 11011;
    INDEXEDDB_ABORT_ERROR: 11012;
    INDEXEDDB_UNKNOWN_ERROR: 11013;
  }

  interface MediaPlayerEvents {
    AST_IN_FUTURE: 'astInFuture';
    BUFFER_EMPTY: 'bufferStalled';
    BUFFER_LOADED: 'bufferLoaded';
    BUFFER_LEVEL_STATE_CHANGED: 'bufferStateChanged';
    BUFFER_LEVEL_UPDATED: 'bufferLevelUpdated';
    CAN_PLAY: 'canPlay';
    CAN_PLAY_THROUGH: 'canPlayThrough';
    CAPTION_RENDERED: 'captionRendered';
    CAPTION_CONTAINER_RESIZE: 'captionContainerResize';
    CONFORMANCE_VIOLATION: 'conformanceViolation';
    DYNAMIC_TO_STATIC: 'dynamicToStatic';
    ERROR: 'error';
    EVENT_MODE_ON_RECEIVE: 'eventModeOnReceive';
    EVENT_MODE_ON_START: 'eventModeOnStart';
    FRAGMENT_LOADING_COMPLETED: 'fragmentLoadingCompleted';
    FRAGMENT_LOADING_PROGRESS: 'fragmentLoadingProgress';
    FRAGMENT_LOADING_STARTED: 'fragmentLoadingStarted';
    FRAGMENT_LOADING_ABANDONED: 'fragmentLoadingAbandoned';
    KEY_ADDED: 'public_keyAdded';
    KEY_ERROR: 'public_keyError';
    KEY_MESSAGE: 'public_keyMessage';
    KEY_SESSION_CLOSED: 'public_keySessionClosed';
    KEY_SESSION_CREATED: 'public_keySessionCreated';
    KEY_SESSION_REMOVED: 'public_keySessionRemoved';
    KEY_STATUSES_CHANGED: 'public_keyStatusesChanged';
    KEY_SYSTEM_SELECTED: 'public_keySystemSelected';
    KEY_SYSTEM_ACCESS_COMPLETE: 'public_keySystemAccessComplete';
    KEY_SESSION_UPDATED: 'public_keySessionUpdated';
    LICENSE_REQUEST_COMPLETE: 'public_licenseRequestComplete';
    LICENSE_REQUEST_SENDING: 'public_licenseRequestSending';
    LOG: 'log';
    MANIFEST_LOADED: 'manifestLoaded';
    MANIFEST_VALIDITY_CHANGED: 'manifestValidityChanged';
    METRICS_CHANGED: 'metricsChanged';
    METRIC_ADDED: 'metricAdded';
    METRIC_CHANGED: 'metricChanged';
    METRIC_UPDATED: 'metricUpdated';
    OFFLINE_RECORD_FINISHED: 'public_offlineRecordFinished';
    OFFLINE_RECORD_LOADEDMETADATA: 'public_offlineRecordLoadedmetadata';
    OFFLINE_RECORD_STARTED: 'public_offlineRecordStarted';
    OFFLINE_RECORD_STOPPED: 'public_offlineRecordStopped';
    PERIOD_SWITCH_STARTED: 'periodSwitchStarted';
    PERIOD_SWITCH_COMPLETED: 'periodSwitchCompleted';
    ADAPTATION_SET_REMOVED_NO_CAPABILITIES: 'adaptationSetRemovedNoCapabilities';
    PLAYBACK_ENDED: 'playbackEnded';
    PLAYBACK_ERROR: 'playbackError';
    PLAYBACK_LOADED_DATA: 'playbackLoadedData';
    PLAYBACK_METADATA_LOADED: 'playbackMetaDataLoaded';
    PLAYBACK_NOT_ALLOWED: 'playbackNotAllowed';
    PLAYBACK_PAUSED: 'playbackPaused';
    PLAYBACK_PLAYING: 'playbackPlaying';
    PLAYBACK_PROGRESS: 'playbackProgress';
    PLAYBACK_RATE_CHANGED: 'playbackRateChanged';
    PLAYBACK_SEEKED: 'playbackSeeked';
    PLAYBACK_SEEKING: 'playbackSeeking';
    PLAYBACK_STALLED: 'playbackStalled';
    PLAYBACK_STARTED: 'playbackStarted';
    PLAYBACK_TIME_UPDATED: 'playbackTimeUpdated';
    PLAYBACK_VOLUME_CHANGED: 'playbackVolumeChanged';
    PLAYBACK_WAITING: 'playbackWaiting';
    PROTECTION_CREATED: 'public_protectioncreated';
    PROTECTION_DESTROYED: 'public_protectiondestroyed';
    REPRESENTATION_SWITCH: 'representationSwitch';
    TRACK_CHANGE_RENDERED: 'trackChangeRendered';
    QUALITY_CHANGE_RENDERED: 'qualityChangeRendered';
    QUALITY_CHANGE_REQUESTED: 'qualityChangeRequested';
    STREAM_ACTIVATED: 'streamActivated';
    STREAM_DEACTIVATED: 'streamDeactivated';
    STREAM_INITIALIZED: 'streamInitialized';
    STREAM_INITIALIZING: 'streamInitializing';
    STREAM_TEARDOWN_COMPLETE: 'streamTeardownComplete';
    STREAM_UPDATED: 'streamUpdated';
    TEXT_TRACKS_ADDED: 'allTextTracksAdded';
    TEXT_TRACK_ADDED: 'textTrackAdded';
    TTML_PARSED: 'ttmlParsed';
    TTML_TO_PARSE: 'ttmlToParse';
  }
// ...
  export class BitrateInfo {
    mediaType: 'video' | 'audio' | 'image';
    bitrate: number;
    width: number;
    height: number;
    scanType: string;
    qualityIndex: number;
  }
// ...
  export class StreamInfo {
    id: string;
    index: number;
    start: number;
    duration: number;
    manifestInfo: IManifestInfo;
    isLast: boolean;
  }
// ...
}